RealSense kernel patches for RPi kernel 5.10.17
This is a combination of 8 commits.

This is the 1st commit message:

Debug ioctl fmt desc

The patch is useful to profile kernel calls invocation after applying kernel patches

This is the commit message #2:

Backporting hotfix patch

The patch is a backporting of Hans Verkuil proposal to circumvent
video buffer de/allocation sync issues by preventing buffers queue operation
during v4l2_streamoff. Without the patch during streamoff operation
the buffers queue continues to change that results
https://patchwork.kernel.org/patch/11095737/

This is the commit message #3:

RealSense camera formats

This is the commit message #4:

RealSense UVC metadata attributes

This is the commit message #5:

Adding missing HID timestamp patch for Gyro sensor.

A symmetric patch for Accelerator was already upstreamed, this patch was
skipped from upstream

The patch was written by Srinivas Pandruvada
<srinivas.pandruvada@intel.com>

Additionally, timestamp conversion is fixed

Signed-off-by: Evgeni <evgeni.raikhel@intel.com>

This is the commit message #6:

Bug fix for incomplete PowerLineFrequency enumeration map

Signed-off-by: Evgeni Raikhel <evgeni.raikhel@intel.com>

This is the commit message #7:

Change UVC_URBS 5->16

This is the commit message #8:

Register format description with ioctls